+++ psycop/projects/cancer/feature_generation/main.py
@@ -1,99 +1,31 @@
 """Main feature generation."""
-import logging
 import sys
 from pathlib import Path
 
-from psycop.common.feature_generation.application_modules.describe_flattened_dataset import (
-    save_flattened_dataset_description_to_disk,
+from psycop.common.feature_generation.application_modules.generate_feature_set import (
+    init_wandb_and_generate_feature_set,
 )
-from psycop.common.feature_generation.application_modules.flatten_dataset import (
-    create_flattened_dataset,
+from psycop.projects.cancer.cancer_config import (
+    get_cancer_feature_specifications,
+    get_cancer_project_info,
 )
-from psycop.common.feature_generation.application_modules.loggers import (
-    init_root_logger,
+from psycop.projects.cancer.feature_generation.cohort_definition.cancer_cohort_definer import (
+    CancerCohortDefiner,
 )
-from psycop.common.feature_generation.application_modules.project_setup import (
-    ProjectInfo,
-    init_wandb,
-)
-from psycop.common.feature_generation.application_modules.save_dataset_to_disk import (
-    split_and_save_dataset_to_disk,
-)
-from psycop.common.feature_generation.application_modules.wandb_utils import (
-    wandb_alert_on_exception,
-)
-from psycop.common.feature_generation.loaders.raw.load_moves import (
-    load_move_into_rm_for_exclusion,
-)
-from psycop.common.feature_generation.loaders.raw.load_visits import (
-    physical_visits_to_psychiatry,
-)
-from psycop.common.global_utils.paths import OVARTACI_SHARED_DIR
-from psycop.projects.cancer.feature_generation.specify_features import FeatureSpecifier
 
-log = logging.getLogger()
-
-
-@wandb_alert_on_exception
-def _generate_feature_set(project_info: ProjectInfo) -> Path:
-    """Main function for loading, generating and evaluating a flattened
-    dataset."""
-    feature_specs = FeatureSpecifier(
-        project_info=project_info,
-        min_set_for_debug=True,  # Remember to set to False when generating full dataset
-    ).get_feature_specs()
-
-    flattened_df = create_flattened_dataset(
-        feature_specs=feature_specs,
-        prediction_times_df=physical_visits_to_psychiatry(timestamps_only=True),
-        drop_pred_times_with_insufficient_look_distance=False,
-        project_info=project_info,
-        quarantine_df=load_move_into_rm_for_exclusion(),
-        quarantine_days=720,
-    )
-
-    split_and_save_dataset_to_disk(
-        flattened_df=flattened_df,
-        project_info=project_info,
-    )
-
-    save_flattened_dataset_description_to_disk(
-        project_info=project_info,
-        feature_specs=feature_specs,  # type: ignore
-    )
-
-    return project_info.flattened_dataset_dir
-
-
-def generate_feature_set() -> Path:
-    # Run elements that are required before wandb init first,
-    # then run the rest in main so you can wrap it all in
-    # wandb_alert_on_exception, which will send a slack alert
-    # if you have wandb alerts set up in wandb
-    project_info = ProjectInfo(
-        project_name="cancer",
-        project_path=OVARTACI_SHARED_DIR / "cancer",
-    )
-
-    init_root_logger(project_info=project_info)
-
-    log.info(f"Stdout level is {logging.getLevelName(log.level)}")
-    log.debug("Debugging is still captured in the log file")
-
-    # Use wandb to keep track of your dataset generations
-    # Makes it easier to find paths on wandb, as well as
-    # allows monitoring and automatic slack alert on failure
+if __name__ == "__main__":
+    # will not run without this chunk, should be fixed properly at some point
     if sys.platform == "win32":
         (Path(__file__).resolve().parents[0] / "wandb" / "debug-cli.onerm").mkdir(
             exist_ok=True,
             parents=True,
         )
-    init_wandb(
-        project_info=project_info,
+
+    init_wandb_and_generate_feature_set(
+        project_info=get_cancer_project_info(),
+        eligible_prediction_times=CancerCohortDefiner.get_filtered_prediction_times_bundle().prediction_times.to_pandas(),
+        feature_specs=get_cancer_feature_specifications(),
+        generate_in_chunks=True,
+        chunksize=10,
+        feature_set_name="with_sentence_transformer",
     )
-
-    return _generate_feature_set(project_info=project_info)
-
-
-if __name__ == "__main__":
-    generate_feature_set()