+++ CHANGELOG.md
@@ -2,6 +2,111 @@
 
 <!--next-version-placeholder-->
 
+## v0.100.0 (2023-08-22)
+
+### Feature
+
+* Filename check earlier for feature-gen ([`17b3404`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/17b3404a3721167f6efad037855655bed8b45831))
+
+## v0.99.0 (2023-08-22)
+
+### Feature
+
+* Cohort creation for the cancer project ([`4a408b9`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/4a408b9f3ccbea0d28acb980a400d5e6b0d8890e))
+
+### Fix
+
+* Correct type hints for aggregation ([`ac9fc29`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/ac9fc29172cca3081fe76f9e640c3689f086a9d7))
+* Reconfigure lab tests ([`caebc33`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/caebc338019f0f05123f38453249ba33a6fb8856))
+* Replaced unsued function ([`75f241c`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/75f241cafff64b55eace870ed73791982d2baa3f))
+
+### Documentation
+
+* Docstring ([`e7a6879`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/e7a6879cf48fcc0ea0d10d2663dfac1e9a78ae2d))
+* Docstring update ([`c98bd22`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/c98bd22ce279d7dab440918772b0312f83612bd4))
+* Update docstring ([`87c5c89`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/87c5c89bcf9cad637e37d4059f63be4707e3d039))
+
+## v0.98.0 (2023-08-22)
+
+### Feature
+
+* Add sentence transformers features ([`8a91048`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/8a9104865f68da1fcf74af37ba0f338e8a3d5b9d))
+* Adding text specs ([`27b37ac`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/27b37ac22c9dddbaa422c67a20afa21610d77373))
+
+### Fix
+
+* Type annotation ([`a18cc64`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/a18cc64996afa1fac283ac1b3388aebd1b54ff34))
+* Wandb back to offline ([`b67d4f7`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/b67d4f7bc33ec4cf2f0fe1f24c804707816a8c85))
+* Config ([`f5e2c73`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/f5e2c738c2fc8bbd592ec1f90902786a4adfbccf))
+* Wandb config ([`761ea6a`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/761ea6a16b5a333c0c7491035323f2653edf10ae))
+* Lookbehind combi config ([`09919f4`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/09919f470efe681a8c305ef0e30f90e8fe549bbc))
+* Update configs ([`6ad6fef`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/6ad6fef3f8b1d8b258c15d4a165c942612e4593a))
+* Correct type hints for aggregation ([`d6f4311`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/d6f4311a34b34a0deebd8ce2d10bc0bd683945cf))
+* Param changes ([`650912d`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/650912df0ea1e2ef276e01fc2da1187696981417))
+* Add missing arg ([`500be32`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/500be3288b0f382403f71f924bd0ad3698ce3645))
+* Minor changes to params ([`dd1dcea`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/dd1dcea57e89baee88390ddaa37b7c17071de74b))
+
+## v0.97.0 (2023-08-15)
+
+### Feature
+
+* Add new dir param and user prompt ([`fe46dbf`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/fe46dbfc528fc83e00fdb371a70d2436c3fb4278))
+
+### Fix
+
+* Broken tests due to missing arg ([`74cd065`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/74cd0651711f0292b11c9dfe84f292e86bd3a119))
+* Add arg to general function ([`6395109`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/639510928f28221c8c0f0068efb90846b0c49a4d))
+* Update general function ([`81506ee`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/81506ee857188b1df32ac785b12d422f42701732))
+* Instructions in README.md ([`c2de501`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/c2de50141da8bb0a6e85fe522ec8e2cb66687687))
+
+## v0.96.0 (2023-08-11)
+
+### Feature
+
+* Encode tfidf values ([`14806f8`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/14806f8ddaa6bccc1598ea0d4a3f559cc7f0b326))
+* Embed text with tfidf model ([`d3c8c25`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/d3c8c255b4b0ce6691337c9e06ce4a36ccf98f48))
+
+### Fix
+
+* Ignore type check ([`17a3b45`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/17a3b45052dde0f5b6e94645d7c87b15d743ad0a))
+
+## v0.95.0 (2023-08-10)
+
+### Feature
+
+* First stab at chunked feature gen ([`0a31a61`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/0a31a613282d87927d4117992c6b9f46e7ee9c2c))
+* Add loader for embedded text ([`40c8271`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/40c8271bd7b45d5c9fbb68b661fd21f385a0aad3))
+* Train sentence transformer code ([`055a572`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/055a572e0151b6044db9a12f491b8d308307b281))
+* Sentence transformer embedding ready to train ([`40ce08a`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/40ce08ab8baa429a7529506938b02015dcf7c50d))
+* Sentence transformer embedding ([`9b998f8`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/9b998f8cfd71e79a33cdd08024a45314923ee345))
+* Vis qc ([`ec8ff15`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/ec8ff15dc22524137d2a0f4f753377c288a172c5))
+
+### Fix
+
+* Misc ([`df49d9b`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/df49d9b3772daaf5017c560abfbdb267da4c8e22))
+* Ignore old import erros ([`d6a2105`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/d6a210572a4d48abb8bfec4396f0668dee19e2f9))
+* Reinstate 'prefixes_to_describe' param ([`761c6f0`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/761c6f06ca34ec3b3f8ea7ee13f522285974e158))
+* Remove old param ([`4240440`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/424044086c361391fe2d93c9e6c524df466843e2))
+* Minor changes and typos ([`2341abd`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/2341abdec245c7558a961d53b3ba2a98374a2bbd))
+* Typo in requirements ([`d97aefe`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/d97aefec0c37af65e26903313c274af091df2ee7))
+* Text feat specs resolve mltp to mean ([`9dd1065`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/9dd1065e10d22a5f5db9eda2d1527b77549274f9))
+* Paths ([`b008a95`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/b008a955c6d900ac089b6b85e9ca9295e86c16a6))
+* Change chunking pipeline ([`781d442`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/781d44229c450eabfcfc462a34ac5a6a6eac5c98))
+* Updating scz_bp feature gen ([`8b1ee14`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/8b1ee142693cdf15faaa0540fba878f0cc894d7e))
+* Move chunk tests ([`df7834d`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/df7834da0494684d0da6eaae9d15df243cd9c1da))
+* Move chunk tests ([`295e4c0`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/295e4c045c06375ebe464db376bc6607bfcbec65))
+* Don't modify prediction_times_df in PredictionTimeFilterer ([`e1eae8c`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/e1eae8c96368d69346feb2cbccf6891af595e536))
+* Type hint for ColNames ([`0454b42`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/0454b4265144b6749da2d35bfa8d144484ae1710))
+* Chunked feature gen ([`d8e2ac7`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/d8e2ac772ac583600d82692dfca56e73a895cfac))
+* Set wandb to offline during feature gen ([`5ce32be`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/5ce32beea252c68dac9b6d7e454f1bc57141ef5e))
+* Print time taken for sentence embedding ([`ac4e8a2`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/ac4e8a2a76cd87a70db3677055d24236188e8106))
+
+## v0.94.0 (2023-08-03)
+
+### Feature
+
+* Multilabel classification ([`5285b2c`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/5285b2c00f63a74cb19b6051903fd85a0df75773))
+
 ## v0.93.0 (2023-07-20)
 
 ### Feature